Fixed a bug in the taylor expansion code
Added in a call to reduce_full before taylor expanding polynomials.
The code makes some assumptions about the structure of a polynomial, so
it must be reduced before the taylor expansion can be computed.

Also added in a new function for parsing polynomials with floating
point coefficients. The old parsing function works, but sometimes
requires the use to specify the type of the polynomial. The new
function specifies the type for you.